Explicitly include header files.
The lgtm.com analysis seems to be *.cc focused. I'm not sure adding the headers helps at all, but the lgtm.com analysis
logs indicate that the YgorClustering headers are indeed being inspected. It might be helpful to adjust the include
location to point to the uninstalled src/, but this seems brittle.