Updated example #4: clustering files based on mod time.
The example now allows the root directory to be specified on the command line. It also ensures the output directory is
present, and prints diagnostic information about which files are being clustered.

The default DBSCAN parameters were modified to provide better out-of-the-box support for clustering photos and videos
for day trips.